Extract supervisor test store support
Move the in-memory session store harness into a dedicated test submodule and describe generation convergence accurately.


Copilot-Session: c4de40cc-4f25-46fb-8cc0-e7096c818c11
